refactor(billing): extract einvoice-types leaf module + drop unsafe cast in dispatcher
