+++ website/src/lib/xrechnung-ubl.ts
@@ -1,4 +1,4 @@
-import type { EInvoiceInput } from './einvoice-profile';
+import type { EInvoiceInput } from './einvoice-types';
 
 export function generateXRechnungUbl(_p: EInvoiceInput): string {
   throw new Error('not implemented'); // siehe Task 5
