+++ website/src/lib/zugferd.ts
@@ -1,3 +1,5 @@
+import type { EInvoiceInput } from './einvoice-types';
+
 // Local type — only the fields used in generateZugferdXml
 interface FullInvoice {
   number: string;
@@ -177,8 +179,6 @@ export function generateZugferdXml(inv: FullInvoice, seller: ZugferdSellerConfig
   </rsm:SupplyChainTradeTransaction>
 </rsm:CrossIndustryInvoice>`;
 }
-
-import type { EInvoiceInput } from './einvoice-profile';
 
 export function generateXRechnungCii(_p: EInvoiceInput): string {
   throw new Error('not implemented'); // siehe Task 4
